Core: Force floating point number
The functions toRadians() or toDegrees() require a floating point type as argument to avoid unexpected roundings when using an integral type

+++ tests/src/Base/Tools.cpp
@@ -39,7 +39,7 @@ TEST(Tools, TestSignum)
 TEST(Tools, TestRadian)
 {
     using Base::numbers::pi;
-    EXPECT_EQ(Base::toRadians<int>(90), 1);
+    EXPECT_EQ(static_cast<int>(Base::toRadians<double>(90)), 1);
     EXPECT_DOUBLE_EQ(Base::toRadians<double>(180), pi);
     EXPECT_DOUBLE_EQ(Base::toRadians<double>(90.0), pi / 2.0);
     EXPECT_DOUBLE_EQ(Base::toRadians<double>(0.0), 0.0);
@@ -48,7 +48,7 @@ TEST(Tools, TestRadian)
 TEST(Tools, TestDegree)
 {
     using Base::numbers::pi;
-    EXPECT_EQ(Base::toDegrees<int>(3), 171);
+    EXPECT_EQ(static_cast<int>(Base::toDegrees<double>(3)), 171);
     EXPECT_DOUBLE_EQ(Base::toDegrees<double>(pi), 180.0);
     EXPECT_DOUBLE_EQ(Base::toDegrees<double>(pi / 2.0), 90.0);
     EXPECT_DOUBLE_EQ(Base::toDegrees<double>(0.0), 0.0);
